Revert "Rename file according to rubocop style"
This reverts commit 9a5eef9804071fd02b8cdde5c89074d43be45358.
