+++ lib/yard-apptweak-template/patches/c_base_handler.rb
