+++ lib/yard-apptweak-template/stubs/autoload.rb
