+++ lib/yard-apptweak-template/templates/apis/fulldoc/text/setup.rb
