+++ lib/yard-apptweak-template/templates/default/docstring/setup.rb
