+++ lib/yard-apptweak-template/templates/default/fulldoc/html/css/sketchup.css
