+++ lib/yard-apptweak-template/templates/default/fulldoc/html/favicon.ico
