+++ lib/yard-apptweak-template/templates/default/fulldoc/html/full_list_object_types.erb
