+++ lib/yard-apptweak-template/templates/default/fulldoc/html/js/jquery-migrate.js
