+++ lib/yard-apptweak-template/templates/default/fulldoc/html/js/sketchup.js
