+++ lib/yard-apptweak-template/templates/default/fulldoc/html/setup.rb
