+++ lib/yard-apptweak-template/templates/default/layout/html/headers.erb
