+++ lib/yard-apptweak-template/templates/default/layout/html/navbar.erb
