+++ lib/yard-apptweak-template/templates/default/layout/html/setup.rb
