+++ lib/yard-apptweak-template/templates/default/method_details/html/method_signature.erb
