+++ lib/yard-apptweak-template/templates/default/method_details/setup.rb
