+++ lib/yard-apptweak-template/templates/default/module/html/box_info.erb
