+++ lib/yard-apptweak-template/templates/default/module/html/constant_summary.erb
