+++ lib/yard-apptweak-template/templates/default/module/html/method_summary.erb
