+++ lib/yard-apptweak-template/templates/inheritance/fulldoc/text/setup.rb
