+++ lib/yard-apptweak-template/templates/rubocop-changelog/fulldoc/text/setup.rb
