+++ lib/yard-apptweak-template/templates/stubs/fulldoc/text/setup.rb
