+++ lib/yard-apptweak-template/templates/versions/fulldoc/text/setup.rb
