+++ lib/yard-apptweak-template/yard/handlers/class_constants.rb
