+++ lib/yard-apptweak-template/yard/handlers/class_enum_constants.rb
