+++ lib/yard-apptweak-template/yard/handlers/global_constants.rb
